experimentscode/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
backports.ssl_match_hostname==3.7.0.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
DataProperty==1.1.0
datasets==3.1.0
dill==0.3.8
docker==7.1.0
exceptiongroup==1.2.2
filelock==3.17.0
flwr==1.16.0
flwr-datasets==0.5.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
grpcio==1.70.0
huggingface-hub==0.28.1
humanreadable==0.4.0
idna==3.10
iniconfig==2.0.0
iterators==0.0.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mbstrdecoder==1.1.4
mdurl==0.1.2
mpmath==1.3.0
msgfy==0.2.1
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
numpy==2.2.3
packaging==24.2
pandas==2.2.3
path==16.16.0
pathspec==0.12.1
pathvalidate==3.2.3
pexpect==4.9.0
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.1
pyparsing==3.2.1
pyroute2==0.8.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-iptables==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
ray==2.42.1
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
seaborn==0.13.2
setuptools==70.2.0
shellingham==1.5.4
SimpleSQLite==1.5.3
six==1.17.0
sqliteschema==2.0.0
subprocrunner==2.0.1
sympy==1.13.1
tabledata==1.3.4
tcconfig==0.29.1
tomli==2.2.1
tomli_w==1.2.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typepy==1.3.4
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
voluptuous==0.15.2
xxhash==3.5.0
yarl==1.18.3
